requirements.txt
boto3==1.17.1
aws_kinesis_agg==1.1.7
gin-config==0.4.0